arm64: Workaround to shorten boot time on the Pinebook Pro
Initialization of crypto_rsa and the tracing subsystem take forever (>300 s)
during early boot. In the case of crypto_rsa it seems to come down to
some misscalculation of crypto test runtime based on dmips specified
in the dt. Loading the module later is way faster (0.5 s). The only
drawback of this is that signing of the mac80211 regdb must also be
disabled.
I do currently not have any idea why initialization of the tracing framework
is taking such a long time (30 s).
However, this workaround is very much non-invasive since it concernes only
the kernel config.
This patch shortens the boot time from more than 3 minutes down to
just 30 seconds even from an micro SD card.