media: rkvdec: Add the rkvdec driver
MRFIXIT: Most of the linuxtv rkvdec branch have been merged into torvalds master, but the rkvdec and support has not, so pull the following **series** of patches from patchwork.kernel.org (links to 1 of X in each series):
https://patchwork.kernel.org/patch/11473825/

Boris Brezillon <boris.brezillon@collabora.com>

The rockchip vdec block is a stateless decoder that's able to decode
H264, HEVC and VP9 content. This commit adds the core infrastructure
and the H264 backend. Support for VP9 and HEVS will be added later on.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Tested-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>