arm64: dts: rockchip: rk3399: Define the rockchip Video Decoder node
MRFIXIT: Most of the linuxtv rkvdec branch have been merged into torvalds master, but the rkvdec and support has not, so pull the following **series** of patches from patchwork.kernel.org (links to 1 of X in each series):
https://patchwork.kernel.org/patch/11473825/

Boris Brezillon <boris.brezillon@collabora.com>

RK3399 has a Video decoder, define the node in the dtsi. We also add
the missing power-domain in mmu node and enable the block.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>