fixup the rock64 DTS
Note that the GPIO-LEDs are configured very wrong - but this was the only way they would work. Heartbeat trigger still doesn't work at all, LED just stays on, and the GPIOs should be LOW and not HIGH. Doesn't make sense why it worked this way but it was the only combination of configs that allowed them to even remotely work correctly.