fixup the rockpro DTS
Note that the GPIO-LEDs are configured very wrong - but this was the only way they would work. Heartbeat trigger still doesn't work at all, LED just stays on, and the GPIOs should be HIGH and not LOW. It also seems like the triggers are applying to opposite LED's maybe. Doesn't make sense why it worked this way but it was the only combination of configs that allowed them to even remotely work correctly.